Implement generic parsing of complex structures
Implement the core algorithm of the whole program: a generic function that can parse complex structures.

In the following commits, I will be implementing parse_x functions where x is a complex structure. They will all just be wrapper functions that pass in the right set of arguments to parse_complex_structure (in other words, parse_complex_structure does all the "heavy lifting")

Note the recursion allows us to avoid keeping track of where we are in the hierarchy right now (i.e the law_dict variable in the iterative version of the algorithm)