[Tree Migration 2/n] Mutate argument > return value
Broad Sketch
---------------
Previously, parse_x functions would return the parsed structure

def parse_x(...):
       child = parse_y(...)
       return ComplexNode(type=Structure.X, children=[child])

This diff migrates to a model where parse_x functions take in the parent node, and then mutates the parent node by adding the parsed structure to the parent's children array

def parse_x(parent, ...):
    x_node = ComplexNode(...)
    parent.add_child(x_node)

    parse_y(x_node, ...) # y is x's child

Why
---------
The new approach gives us access to a snapshot of the whole parsing tree at any given moment in time

e.g if we need to crash, we can print out a snapshot of what we have parsed so far

e.g if we reach an area of ambiguity where we want to hand over control to a human, we can print out a snapshot of what we have parsed so far alongside the problematic line

e.g if we reach what may be an embedded pasal, we we can traverse the current snapshot for clues by understanding the whole hierarchy of the tree

This was not possible in the previous approach. By design, the previous approach's emphasis on passing parsed structures via return values & avoiding mutation of arguments provides a clean encapsulation between functions: at any point we have no awareness of the global tree; each function is just focused on parsing its own subtree. I sitll think this approach is cleaner/more elegant, but is not as useful for reasons stated above